refactor: move rewind resources to willo module
