+++ src/main.rs
@@ -98,7 +98,6 @@ fn main() {
         .insert_resource(Msaa { samples: 1 })
         .insert_resource(level_selection.clone())
         .insert_resource(level_transition::TransitionTo(level_selection))
-        .insert_resource(resources::RewindSettings::NORMAL)
         .add_startup_system(gameplay::transitions::spawn_ui_root)
         .add_system_to_stage(CoreStage::PreUpdate, sugar::make_ui_visible)
         .add_system(
