+++ src/willo.rs
@@ -6,14 +6,14 @@ use crate::{
     history::FlushHistoryCommands,
     history::{History, HistoryCommands},
     movement_table::Direction,
-    resources::{RewindSettings, RewindTimer},
     sokoban::RigidBody,
     AssetHolder, GameState, SystemLabels,
 };
-use bevy::{prelude::*, utils::Duration};
+use bevy::prelude::*;
 use bevy_easings::*;
 use bevy_ecs_ldtk::prelude::*;
 use iyes_loopless::prelude::*;
+use std::{ops::Range, time::Duration};
 
 /// Plugin providing functionality for Willo, the player character.
 pub struct WilloPlugin;
@@ -21,6 +21,7 @@ pub struct WilloPlugin;
 impl Plugin for WilloPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugin(FromComponentAnimator::<WilloAnimationState>::new())
+            .init_resource::<RewindSettings>()
             .add_event::<WilloMovementEvent>()
             .add_system(
                 willo_input
@@ -130,6 +131,38 @@ pub struct MovementTimer(pub Timer);
 impl Default for MovementTimer {
     fn default() -> MovementTimer {
         MovementTimer(Timer::from_seconds(MOVEMENT_SECONDS, false))
+    }
+}
+
+#[derive(Clone, Debug, Default)]
+pub struct RewindTimer {
+    pub velocity: f32,
+    pub timer: Timer,
+}
+
+impl RewindTimer {
+    pub fn new(millis: u64) -> RewindTimer {
+        RewindTimer {
+            velocity: millis as f32,
+            timer: Timer::new(Duration::from_millis(millis), true),
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct RewindSettings {
+    pub hold_range_millis: Range<u64>,
+    pub hold_acceleration: f32,
+    pub hold_timer: Option<RewindTimer>,
+}
+
+impl Default for RewindSettings {
+    fn default() -> Self {
+        RewindSettings {
+            hold_range_millis: 50..200,
+            hold_acceleration: 50.,
+            hold_timer: None,
+        }
     }
 }
